requirements.txt
dash
dash-bootstrap-components
pandas
requests
pillow
wheel